# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SAT Tutor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Question {
  id: string;
  subject: string;
  /** @nullable */
  topic?: string | null;
  /** @nullable */
  difficulty?: string | null;
  /** @nullable */
  passage?: string | null;
  question: string;
  options: string[];
  correct: string;
  /** @nullable */
  explanation?: string | null;
}

export interface ExplanationInput {
  question: string;
  /** @nullable */
  passage?: string | null;
  options: string[];
  correct: string;
  userAnswer: string;
  subject?: string;
}

export interface Explanation {
  text: string;
}

export interface QuestionSnapshot {
  question?: string;
  /** @nullable */
  passage?: string | null;
  options?: string[];
  /** @nullable */
  explanation?: string | null;
}

export interface ProgressEntry {
  id: string;
  timestamp: string;
  questionId: string;
  topic: string;
  userAnswer: string;
  correct: string;
  isCorrect: boolean;
  subject: string;
  snapshot?: QuestionSnapshot;
}

export interface ProgressInput {
  questionId: string;
  topic: string;
  userAnswer: string;
  correct: string;
  isCorrect: boolean;
  subject: string;
  snapshot?: QuestionSnapshot;
}

export interface TopicStat {
  topic: string;
  correct: number;
  incorrect: number;
  total: number;
  accuracy: number;
}

export interface DbStatus {
  loaded: boolean;
  englishCount: number;
  mathCount: number;
  /** @nullable */
  message?: string | null;
}

export type GetQuestionsParams = {
subject: GetQuestionsSubject;
difficulty?: GetQuestionsDifficulty;
count?: number;
/**
 * Comma-separated list of mastered question IDs to exclude
 */
masteredIds?: string;
};

export type GetQuestionsSubject = typeof GetQuestionsSubject[keyof typeof GetQuestionsSubject];


export const GetQuestionsSubject = {
  english: 'english',
  math: 'math',
} as const;

export type GetQuestionsDifficulty = typeof GetQuestionsDifficulty[keyof typeof GetQuestionsDifficulty];


export const GetQuestionsDifficulty = {
  Easy: 'Easy',
  Medium: 'Medium',
  Hard: 'Hard',
} as const;

export type GetProgressParams = {
subject: GetProgressSubject;
};

export type GetProgressSubject = typeof GetProgressSubject[keyof typeof GetProgressSubject];


export const GetProgressSubject = {
  english: 'english',
  math: 'math',
} as const;

export type ResetProgressParams = {
subject: ResetProgressSubject;
};

export type ResetProgressSubject = typeof ResetProgressSubject[keyof typeof ResetProgressSubject];


export const ResetProgressSubject = {
  english: 'english',
  math: 'math',
} as const;

export type GetStatsParams = {
subject: GetStatsSubject;
};

export type GetStatsSubject = typeof GetStatsSubject[keyof typeof GetStatsSubject];


export const GetStatsSubject = {
  english: 'english',
  math: 'math',
} as const;