# lib/api-zod/src/generated/types/explanationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SAT Tutor API
 * OpenAPI spec version: 0.1.0
 */

export interface ExplanationInput {
  question: string;
  /** @nullable */
  passage?: string | null;
  options: string[];
  correct: string;
  userAnswer: string;
  subject?: string;
}