# lib/api-zod/src/generated/types/progressInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SAT Tutor API
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionSnapshot } from './questionSnapshot';

export interface ProgressInput {
  questionId: string;
  topic: string;
  userAnswer: string;
  correct: string;
  isCorrect: boolean;
  subject: string;
  snapshot?: QuestionSnapshot;
}