# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SAT Tutor API
 * OpenAPI spec version: 0.1.0
 */

export interface Question {
  id: string;
  subject: string;
  /** @nullable */
  topic?: string | null;
  /** @nullable */
  difficulty?: string | null;
  /** @nullable */
  passage?: string | null;
  question: string;
  options: string[];
  correct: string;
  /** @nullable */
  explanation?: string | null;
}