# lib/api-zod/src/generated/types/questionSnapshot.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SAT Tutor API
 * OpenAPI spec version: 0.1.0
 */

export interface QuestionSnapshot {
  question?: string;
  /** @nullable */
  passage?: string | null;
  options?: string[];
  /** @nullable */
  explanation?: string | null;
}